Rename emscripten/math.h to emscripten/em_math.h to avoid possible name collisions if emscripten/ is in -I
